// class-room-sauvegarde/11/class-room-frontend/src/app/api/model/contenuDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LocalTime } from './localTime';
import { SeanceDto } from './seanceDto';


export interface ContenuDto { 
    idContenu?: number;
    date?: string;
    heure?: LocalTime;
    contenu?: string;
    seance?: SeanceDto;
}